requirements.txt
APScheduler==3.10.0
asgiref==3.6.0
async-generator==1.10
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
Django==4.1.6
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.1.0
h11==0.14.0
idna==3.4
Markdown==3.4.1
numpy==1.24.2
outcome==1.2.0
pandas==1.5.3
pycparser==2.21
PyJWT==2.6.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
selenium==4.8.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
trio==0.22.0
trio-websocket==0.9.2
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
wsproto==1.2.0
